+++ th2_data_services/provider/v5/provider_api/http.py
@@ -24,7 +24,8 @@ from urllib.parse import quote
 from th2_data_services.provider.interfaces.source_api import IHTTPProviderSourceAPI
 from th2_data_services.decode_error_handler import UNICODE_REPLACE_HANDLER
 
-logger = logging.getLogger(__name__)
+logger = logging.getLogger("th2_data_services")
+logger.setLevel(logging.DEBUG)
 
 
 class HTTPProvider5API(IHTTPProviderSourceAPI):
@@ -69,7 +70,7 @@ class HTTPProvider5API(IHTTPProviderSourceAPI):
         query = ""
         for id in ids:
             query += f"ids={id}&"
-        return self.__encode_url(f"{self._url}/events/?{query[:-1]}")
+        return self.__encode_url(f"{self._url}/events?{query[:-1]}")
 
     def get_url_find_message_by_id(self, message_id: str) -> str:
         """REST-API `message` call returns a single message with the specified id."""
@@ -153,7 +154,7 @@ class HTTPProvider5API(IHTTPProviderSourceAPI):
         }
 
         query = ""
-        url = f"{self._url}/search/sse/events/?"
+        url = f"{self._url}/search/sse/events?"
         for k, v in kwargs.items():
             if v is None:
                 continue
@@ -196,7 +197,7 @@ class HTTPProvider5API(IHTTPProviderSourceAPI):
         }
 
         query = ""
-        url = f"{self._url}/search/sse/messages/?"
+        url = f"{self._url}/search/sse/messages?"
         for k, v in kwargs.items():
             if v is None:
                 continue